# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/v1/mobile/docs": {
      "get": {
        "operationId": "MobileController_getDocs_v1",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Mobile"
        ]
      }
    },
...
    "/api/v1/v1/study/quiz/generate": {
      "post": {
        "operationId": "QuizController_generateQuiz_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
    "/api/v1/v1/study/quiz/{id}/submit": {
      "post": {
        "operationId": "QuizController_submitQuiz_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
    "/api/v1/v1/study/quiz/{id}": {
      "get": {
        "operationId": "QuizController_getQuiz_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
...
    "/api/v1/v1/documents/upload": {
      "post": {
        "operationId": "DocumentsController_uploadFile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
...
    "/api/v1/v1/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "delete": {
        "operationId": "DocumentsController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/v1/documents/{id}/assets": {
      "get": {
        "operationId": "DocumentsController_getAssets_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/v1/documents/{id}/status": {
      "get": {
        "operationId": "DocumentsController_getStatus_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/v1/documents/{id}/chunks": {
      "get": {
        "operationId": "DocumentsController_getChunks_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/v1/documents/{id}/metadata": {
      "get": {
        "operationId": "DocumentsController_getMetadata_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
...
    "/api/v1/v1/groups/{id}/docs": {
      "post": {
        "operationId": "StudyGroupController_addDocument_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StudyGroup"
        ]
      }
    },
    "/api/v1/v1/groups/{id}/docs/{docId}": {
      "delete": {
        "operationId": "StudyGroupController_removeDocument_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StudyGroup"
        ]
      }
    },
...
    "/api/public/v1/documents": {
      "get": {
        "description": "Retrieve a cursor-paginated list of processed organization documents.",
        "operationId": "PublicDocumentsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents retrieved successfully."
          },
          "401": {
            "description": "Invalid or missing API key."
          },
          "403": {
            "description": "Insufficient scopes."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List public documents",
        "tags": [
          "Public Documents"
        ]
      },
      "post": {
        "description": "Upload a document file to the organization study processing pipeline.",
        "operationId": "PublicDocumentsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded and processing queued successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload document",
        "tags": [
          "Public Documents"
        ]
      }
    },
    "/api/public/v1/documents/{id}": {
      "get": {
        "description": "Retrieve metadata details of a specific organization document.",
        "operationId": "PublicDocumentsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document metadata retrieved successfully."
          },
          "404": {
            "description": "Document not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get document by ID",
        "tags": [
          "Public Documents"
        ]
      },
      "delete": {
        "description": "Permantently delete a document, its chunks, storage objects, and vector embeddings.",
        "operationId": "PublicDocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete document",
        "tags": [
          "Public Documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Study Assistant Public API",
    "description": "Production-grade developer APIs for AI Study Assistant Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your API Key prefixed with Bearer (e.g. Bearer ska_live_...)",
        "in": "header"
      }
    },
    "schemas": {
...
      "GenerateQuizDto": {
        "type": "object",
        "properties": {}
      },
      "SubmitQuizDto": {
        "type": "object",
        "properties": {}
      },
      "UploadDocumentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}